--- a/src/afl-fuzz-run.c
+++ b/src/afl-fuzz-run.c
@@ -367,13 +367,45 @@ abort_calibration:
 
 }
 
+/* this is a memstr() function where the needle has a fixed length */
+
+static u8 *next_entry(u8 *entry, u8 *file_list, u32 file_list_len) {
+
+  register int i;
+
+  for (i = 0; i < file_list_len - 6; i++)
+    if (memcmp(file_list + i, entry, 6) == 0) return (file_list + i);
+
+  return NULL;
+
+}
+
+/* Generate next entry to search for in the saved directory listing.
+   Dirty hack: only use sprintf if the new number ends in a 0, otherwise
+   we can be cheap and just increase the least significant byte */
+static void update_entry(u8 *entry, u32 next_accept) {
+
+  if (next_accept % 10 == 0)
+    sprintf(entry, "%06u", next_accept);
+  else
+    entry[5] += 1;
+
+}
+
 /* Grab interesting test cases from other fuzzers. */
 
 void sync_fuzzers(afl_state_t *afl) {
 
+  u64             profile_start, profile_end;
+  struct timespec spec;
+  clock_gettime(CLOCK_REALTIME, &spec);
+  profile_start = (spec.tv_sec * 1000000000) + spec.tv_nsec;
+
   DIR *          sd;
   struct dirent *sd_ent;
   u32            sync_cnt = 0;
+  u8 *           file_list, path[PATH_MAX], entry[8] = {0};
+  size_t         file_list_size = 65536;
 
   sd = opendir(afl->sync_dir);
   if (!sd) { PFATAL("Unable to open '%s'", afl->sync_dir); }
@@ -384,12 +416,15 @@ void sync_fuzzers(afl_state_t *afl) {
   /* Look at the entries created for every other fuzzer in the sync directory.
    */
 
+  file_list =
+      ck_maybe_grow((void **)&file_list, &file_list_size, file_list_size);
+
   while ((sd_ent = readdir(sd))) {
 
     DIR *          qd;
     struct dirent *qd_ent;
-    u8 *           qd_path, *qd_synced_path;
-    u32            min_accept = 0, next_min_accept;
+    u8 *           qd_path, *qd_synced_path, *next_fn;
+    u32            min_accept = 0, next_accept, file_list_len = 0;
 
     s32 id_fd;
 
@@ -437,8 +472,6 @@ void sync_fuzzers(afl_state_t *afl) {
 
     }
 
-    next_min_accept = min_accept;
-
     /* Show stats */
 
     snprintf(afl->stage_name_buf, STAGE_BUF_SIZE, "sync %u", ++sync_cnt);
@@ -447,45 +480,48 @@ void sync_fuzzers(afl_state_t *afl) {
     afl->stage_cur = 0;
     afl->stage_max = 0;
 
-    /* For every file queued by this fuzzer, parse ID and see if we have
-       looked at it before; exec a test case if not. */
+    /* Read the filelist to memory */
 
     while ((qd_ent = readdir(qd))) {
 
-      u8 *        path;
+      if (qd_ent->d_name[0] == '.') { continue; }
+
+      if (file_list_len + PATH_MAX >= file_list_size)
+        file_list = ck_maybe_grow((void **)&file_list, &file_list_size,
+                                  file_list_size + PATH_MAX + 16384);
+
+      u32 fn_len = strlen(qd_ent->d_name) + 1;  // with null
+
+      memcpy(file_list + file_list_len, qd_ent->d_name, fn_len);
+      file_list_len += fn_len;
+
+    }
+
+    if (!file_list_len) continue;
+
+    next_accept = min_accept;
+    sprintf(entry, "%06u", next_accept);
+
+    /* For every file queued by this fuzzer, parse ID and see if we have
+       looked at it before; exec a test case if not. */
+
+    while ((next_fn = next_entry(entry, file_list, file_list_len))) {
+
       s32         fd;
       struct stat st;
 
-      if (qd_ent->d_name[0] == '.' ||
-          sscanf(qd_ent->d_name, CASE_PREFIX "%06u", &afl->syncing_case) != 1 ||
-          afl->syncing_case < min_accept) {
-
-        continue;
-
-      }
-
       /* OK, sounds like a new one. Let's give it a try. */
 
-      if (afl->syncing_case >= next_min_accept) {
-
-        next_min_accept = afl->syncing_case + 1;
-
-      }
-
-      path = alloc_printf("%s/%s", qd_path, qd_ent->d_name);
+      afl->syncing_case = next_accept;
+      sprintf(path, "%s/%s", qd_path, next_fn);
 
       /* Allow this to fail in case the other fuzzer is resuming or so... */
 
       fd = open(path, O_RDONLY);
 
-      if (fd < 0) {
+      if (fd < 0) { continue; }
 
-        ck_free(path);
-        continue;
-
-      }
-
-      if (fstat(fd, &st)) { PFATAL("fstat() failed"); }
+      if (fstat(fd, &st)) { continue; }
 
       /* Ignore zero-sized or oversized files. */
 
@@ -516,12 +552,13 @@ void sync_fuzzers(afl_state_t *afl) {
 
       }
 
-      ck_free(path);
       close(fd);
+      next_accept++;
+      update_entry(entry, next_accept);
 
     }
 
-    ck_write(id_fd, &next_min_accept, sizeof(u32), qd_synced_path);
+    ck_write(id_fd, &next_accept, sizeof(u32), qd_synced_path);
 
   close_sync:
     close(id_fd);
@@ -531,7 +568,14 @@ void sync_fuzzers(afl_state_t *afl) {
 
   }
 
+  ck_free(file_list);
   closedir(sd);
+
+  clock_gettime(CLOCK_REALTIME, &spec);
+  profile_end = (spec.tv_sec * 1000000000) + spec.tv_nsec;
+  sprintf(path, "echo %016llu >> /tmp/profile.out",
+          profile_end - profile_start);
+  if (system(path) != 0) WARNF("system");
 
 }