This is a new method to sync.

it is slightly faster on intial imports and very slightly slower afterwards.

the advantage: much fewer entries are synced to the queue, because I sort the entries and add them reverse, hence the deeper running queue entries get added first, and shallow ones from the beginning will not.

I think this is good, but I want your opinion on this

Timings - (import queue with 2234 files)

new - 978 imports, 23.58s initial import, 0.00118 per reassessment without adding anything

old - 1113 imports, 24.66s initial import, 0.00096 per reassessment without adding anything

Note that I already made the following changes to dev:
 * only masters sync from everyone, and slaves only sync from a master.
 * when a master starts and detects another master running already a warning is printed and 1s wait
 * when a slave starts and detects no master a warning is printed and 1s wait

What do you think a) about the new algo, and b) about the changes of slaves only syncing from master?