@andreafioraldi 2020-05-17
`only masters sync from everyone, and slaves only sync from a master.`
What about when there is no master?