@vanhauser-thc 2020-05-17
> `only masters sync from everyone, and slaves only sync from a master.`
> What about when there is no master?

"when a slave starts and detects no master a warning is printed and a 5 second wait"

I also updated all documentation to state that one master is needed.