@vanhauser-thc 2020-05-17
yes this can go wrong for people who are used to it without a master. but it hurts performance if everyone syncs from everyone