@andreafioraldi 2020-05-19
Master is useless with structured formats. We should suggest to use -M -d instead of -S for that case.