@andreafioraldi 2020-05-19
Btw there are tools, like QSYM, that run only one instance of AFL and it can not be a master. What about, if not master is found, do a leader election between the slaves? We can just sort the slaves by name and pick the first.