@vanhauser-thc 2020-05-19
> What about, if not master is found, do a leader election between the slaves? We can just sort the slaves by name and pick the first.

That is an option, it then increases complexity. say, what if the master is started later. the newly elected slave would need to see that another master is there now and become a slave again. or if the slave that became a master is terminated, that a different slave picks it up then.

hmmm
but actually it is not that hard to code, and the overhead is limited.