@vanhauser-thc 2020-05-19
> What about if the master dies or crashes half way through fuzzing?

if afl-fuzz crashes then people should create an issue :)