@andreafioraldi 2020-05-20
> > What about if the master dies or crashes half way through fuzzing?
> 
> if afl-fuzz crashes then people should create an issue :)

Probably @domenukk is talking about OOM kill bu out of memory of out of disk space (?)