@andreafioraldi 2020-05-20
Or maybe the user stops one fuzzer at time to use cmin and for some time the master is down