@vanhauser-thc 2020-05-20
if kill -STOP on the master, then it stops. yes.
OOM or disk fill would affect all fuzzers.

I mean of course there are unlikely circumstances or bad usage that can result in problems. but then these are clear problems. IMHO its fine.